+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * Fix `filter_options` being added multiple times and at the incorrect level (#100)
 * Fix missing localization for filter placeholders and page size selector (#92)
+* Fix styling of filter container in DaisyUI theme (#101)
 
 ## v0.9.0-beta.5 (2026-01-27)